# openapi_pretty.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HappyDog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/comments/posts/{post_id}/comments": {
      "post": {
        "summary": "특정 게시글에 새로운 댓글을 작성합니다.",
        "description": "- 성공 시, 생성된 댓글 정보를 201 Created 상태 코드와 함께 반환합니다.\n    - 댓글 생성 후 게시물 작성자 및 멘션된 사용자에게 알림이 생성됩니다.",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "comments"
        ]
      },
      "get": {
        "summary": "특정 게시글의 댓글 목록을 페이지네이션으로 조회합니다.",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'post_id'"
          }
        ],
        "tags": [
          "comments"
        ]
      }
    },
    "/api/comments/comments/{comment_id}": {
      "delete": {
        "summary": "특정 댓글을 삭제합니다. (작성자 본인만 가능)",
        "description": "- 성공 시, 게시물의 댓글 수가 1 감소합니다.",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'comment_id'"
          }
        ],
        "tags": [
          "comments"
        ]
      }
    },
    "/api/comments/comments/{comment_id}/like": {
      "post": {
        "summary": "특정 댓글의 좋아요를 누르거나 취소합니다.",
        "description": "",
        "responses": {
          "200": {
            "description": "성공"
          },
          "400": {
            "$ref": "#/components/responses/VALIDATION_ERROR"
          },
          "401": {
            "$ref": "#/components/responses/UNAUTHORIZED"
          },
          "403": {
            "$ref": "#/components/responses/FORBIDDEN"
          },
          "404": {
            "$ref": "#/components/responses/NOT_FOUND"
          },
          "500": {
            "$ref": "#/components/responses/INTERNAL_ERROR"
          }
        },
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path parameter 'comment_id'"
          }
        ],
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "components": {
...
    "responses": {
      "UNAUTHORIZED": {
        "description": "인증이 필요하거나 토큰이 유효하지 않습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "UNAUTHORIZED"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "VALIDATION_ERROR": {
        "description": "입력 값이 유효하지 않습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "VALIDATION_ERROR"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "FORBIDDEN": {
        "description": "권한이 없습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "FORBIDDEN"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "NOT_FOUND": {
        "description": "리소스를 찾을 수 없습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "NOT_FOUND"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
      "INTERNAL_ERROR": {
        "description": "서버 내부 오류가 발생했습니다.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error_code": {
                  "type": "string",
                  "example": "INTERNAL_ERROR"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "error_code",
                "message"
              ]
            }
          }
        }
      },
...
    }
  },
...
}